Complete logic for the game (for now)
The logic checks whether the user or opponent has 3 in a row
squares marked, and then displays the winner. If the game ends
in a tie, it displays 'Game ended in a tie'

Player factory function is used to get the name, marker, and an array
for both the player and opponent. This is a clean way to give both
their own array, which is then later used to check if any of the
winningIndexes match the player array inedexes

Then the Player function calls Gamelogic functions for both user and
opponent, and the Gamelogic factory function keeps track of the game
progress